backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
unstructured-client = "^0.29.0"
python-dotenv = "^1.0.1"
pymupdf = "^1.25.2"
langchain = "^0.3.15"
langchain-openai = "^0.3.2"
diskcache = "^5.6.3"


python-docx = "^1.1.2"
python-pptx = "^1.0.2"
newsapi-python = "^0.2.7"
langchain-community = "^0.3.15"
aiohttp = "^3.11.11"
fastapi = {extras = ["standard"], version = "^0.115.7"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.20"
sqlalchemy = "^2.0.37"
psycopg2-binary = "^2.9.10"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
alembic = "^1.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
